Deletion of Raster Layer Geopackage sample, renaming Read Geopackage, updated readme's
