+++ src/MAUI/Maui.Samples/Samples/Data/AddRastersAndFeatureTablesFromGeopackage/AddRastersAndFeatureTablesFromGeopackage.xaml
@@ -1,7 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage x:Class="ArcGIS.Samples.RasterLayerGeoPackage.RasterLayerGeoPackage"
+
+<ContentPage x:Class="ArcGIS.Samples.AddRastersAndFeatureTablesFromGeopackage.AddRastersAndFeatureTablesFromGeopackage"
              xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:esriUI="clr-namespace:Esri.ArcGISRuntime.Maui;assembly=Esri.ArcGISRuntime.Maui">
-    <esriUI:MapView x:Name="MyMapView" />
+    <Grid>
+        <esriUI:MapView x:Name="MyMapView" />
+    </Grid>
 </ContentPage>
