--- src/MAUI/Maui.Samples/Samples/Data/ReadGeoPackage/ReadGeoPackage.xaml
@@ -1,10 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-
-<ContentPage x:Class="ArcGIS.Samples.ReadGeoPackage.ReadGeoPackage"
-             xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:esriUI="clr-namespace:Esri.ArcGISRuntime.Maui;assembly=Esri.ArcGISRuntime.Maui">
-    <Grid>
-        <esriUI:MapView x:Name="MyMapView" />
-    </Grid>
-</ContentPage>
