+++ src/WPF/WPF.Viewer/Samples/Data/AddRastersAndFeatureTablesFromGeopackage/AddRastersAndFeatureTablesFromGeopackage.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGIS.WPF.Samples.RasterLayerGeoPackage.RasterLayerGeoPackage"
+﻿<UserControl x:Class="ArcGIS.WPF.Samples.AddRastersAndFeatureTablesFromGeopackage.AddRastersAndFeatureTablesFromGeopackage"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013">
