+++ src/WPF/WPF.Viewer/Samples/Data/AddRastersAndFeatureTablesFromGeopackage/AddRastersAndFeatureTablesFromGeopackage.xaml.cs
@@ -12,20 +12,22 @@ using Esri.ArcGISRuntime.Data;
 using Esri.ArcGISRuntime.Mapping;
 using Esri.ArcGISRuntime.Rasters;
 using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
+using System.Windows;
 
-namespace ArcGIS.WinUI.Samples.ReadGeoPackage
+namespace ArcGIS.WPF.Samples.AddRastersAndFeatureTablesFromGeopackage
 {
     [ArcGIS.Samples.Shared.Attributes.Sample(
-        name: "Read GeoPackage",
+        name: "Add rasters and feature tables from geopackage",
         category: "Data",
         description: "Add rasters and feature tables from a GeoPackage to a map.",
         instructions: "When the sample loads, the feature tables and rasters from the GeoPackage will be shown on the map.",
         tags: new[] { "OGC", "container", "layer", "map", "package", "raster", "table" })]
     [ArcGIS.Samples.Shared.Attributes.OfflineData("68ec42517cdd439e81b036210483e8e7")]
-    public partial class ReadGeoPackage
+    public partial class AddRastersAndFeatureTablesFromGeopackage
     {
-        public ReadGeoPackage()
+        public AddRastersAndFeatureTablesFromGeopackage()
         {
             InitializeComponent();
             _ = Initialize();
@@ -35,18 +37,21 @@ namespace ArcGIS.WinUI.Samples.ReadGeoPackage
         {
             // Create a new map centered on Aurora Colorado.
             MyMapView.Map = new Map(BasemapStyle.ArcGISStreets);
-            MyMapView.Map.InitialViewpoint = new Viewpoint(39.7294, -104.73, 175000);
+            MyMapView.Map.InitialViewpoint = new Viewpoint(39.7294, -104.70, 175000);
 
             // Get the full path to the GeoPackage on the device.
-            string myGeoPackagePath = DataManager.GetDataFolder("68ec42517cdd439e81b036210483e8e7", "AuroraCO.gpkg");
+            string myGeoPackagePath = GetGeoPackagePath();
 
             try
             {
                 // Open the GeoPackage.
                 GeoPackage myGeoPackage = await GeoPackage.OpenAsync(myGeoPackagePath);
 
+                // Get the read only list of GeoPackageRasters from the GeoPackage.
+                IReadOnlyList<GeoPackageRaster> myReadOnlyListOfGeoPackageRasters = myGeoPackage.GeoPackageRasters;
+
                 // Loop through each GeoPackageRaster.
-                foreach (GeoPackageRaster oneGeoPackageRaster in myGeoPackage.GeoPackageRasters)
+                foreach (GeoPackageRaster oneGeoPackageRaster in myReadOnlyListOfGeoPackageRasters)
                 {
                     // Create a RasterLayer from the GeoPackageRaster.
                     RasterLayer myRasterLayer = new RasterLayer(oneGeoPackageRaster)
@@ -55,12 +60,15 @@ namespace ArcGIS.WinUI.Samples.ReadGeoPackage
                         Opacity = 0.55
                     };
 
-                    // Add the raster layer to the map.
+                    // Add the layer to the map.
                     MyMapView.Map.OperationalLayers.Add(myRasterLayer);
                 }
 
+                // Get the read only list of GeoPackageFeatureTables from the GeoPackage.
+                IReadOnlyList<GeoPackageFeatureTable> myReadOnlyListOfGeoPackageFeatureTables = myGeoPackage.GeoPackageFeatureTables;
+
                 // Loop through each GeoPackageFeatureTable.
-                foreach (GeoPackageFeatureTable oneGeoPackageFeatureTable in myGeoPackage.GeoPackageFeatureTables)
+                foreach (GeoPackageFeatureTable oneGeoPackageFeatureTable in myReadOnlyListOfGeoPackageFeatureTables)
                 {
                     // Create a FeatureLayer from the GeoPackageFeatureLayer.
                     FeatureLayer myFeatureLayer = new FeatureLayer(oneGeoPackageFeatureTable);
@@ -71,8 +79,10 @@ namespace ArcGIS.WinUI.Samples.ReadGeoPackage
             }
             catch (Exception e)
             {
-                await new MessageDialog2(e.ToString(), "Error").ShowAsync();
+                MessageBox.Show(e.ToString(), "Error");
             }
         }
+
+        private static string GetGeoPackagePath() => DataManager.GetDataFolder("68ec42517cdd439e81b036210483e8e7", "AuroraCO.gpkg");
     }
 }
