--- src/WPF/WPF.Viewer/Samples/Data/RasterLayerGeoPackage/readme.metadata.json
@@ -1,36 +1,0 @@
-{
-    "category": "Data",
-    "description": "Display a raster contained in a GeoPackage.",
-    "formal_name": "RasterLayerGeoPackage",
-    "ignore": false,
-    "images": [
-        "RasterLayerGeoPackage.jpg"
-    ],
-    "keywords": [
-        "OGC",
-        "container",
-        "data",
-        "image",
-        "import",
-        "layer",
-        "package",
-        "raster",
-        "visualization"
-    ],
-    "offline_data": [
-        "68ec42517cdd439e81b036210483e8e7"
-    ],
-    "redirect_from": [
-        "/net/latest/wpf/sample-code/raster-layer-geopackage.htm"
-    ],
-    "relevant_apis": [
-        "GeoPackage",
-        "GeoPackageRaster",
-        "RasterLayer"
-    ],
-    "snippets": [
-        "RasterLayerGeoPackage.xaml.cs",
-        "RasterLayerGeoPackage.xaml"
-    ],
-    "title": "Raster layer (GeoPackage)"
-}
