--- src/WPF/WPF.Viewer/Samples/Data/ReadGeoPackage/ReadGeoPackage.xaml
@@ -1,8 +1,0 @@
-﻿<UserControl x:Class="ArcGIS.WPF.Samples.ReadGeoPackage.ReadGeoPackage"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013">
-    <Grid>
-        <esri:MapView x:Name="MyMapView" />
-    </Grid>
-</UserControl>
