+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Data/AddRastersAndFeatureTablesFromGeopackage/AddRastersAndFeatureTablesFromGeopackage.jpg
