+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Data/AddRastersAndFeatureTablesFromGeopackage/AddRastersAndFeatureTablesFromGeopackage.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGIS.WinUI.Samples.ReadGeoPackage.ReadGeoPackage"
+﻿<UserControl x:Class="ArcGIS.WinUI.Samples.AddRastersAndFeatureTablesFromGeopackage.AddRastersAndFeatureTablesFromGeopackage"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:esriUI="using:Esri.ArcGISRuntime.UI.Controls">
