+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Data/AddRastersAndFeatureTablesFromGeopackage/AddRastersAndFeatureTablesFromGeopackage.xaml.cs
@@ -1,4 +1,4 @@
-// Copyright 2022 Esri.
+// Copyright 2018 Esri.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at: http://www.apache.org/licenses/LICENSE-2.0
@@ -11,19 +11,21 @@ using ArcGIS.Samples.Managers;
 using Esri.ArcGISRuntime.Data;
 using Esri.ArcGISRuntime.Mapping;
 using Esri.ArcGISRuntime.Rasters;
+using System;
+using System.Threading.Tasks;
 
-namespace ArcGIS.Samples.ReadGeoPackage
+namespace ArcGIS.WinUI.Samples.AddRastersAndFeatureTablesFromGeopackage
 {
     [ArcGIS.Samples.Shared.Attributes.Sample(
-        name: "Read GeoPackage",
+        name: "Add rasters and feature tables from geopackage",
         category: "Data",
         description: "Add rasters and feature tables from a GeoPackage to a map.",
         instructions: "When the sample loads, the feature tables and rasters from the GeoPackage will be shown on the map.",
         tags: new[] { "OGC", "container", "layer", "map", "package", "raster", "table" })]
     [ArcGIS.Samples.Shared.Attributes.OfflineData("68ec42517cdd439e81b036210483e8e7")]
-    public partial class ReadGeoPackage : ContentPage
+    public partial class AddRastersAndFeatureTablesFromGeopackage
     {
-        public ReadGeoPackage()
+        public AddRastersAndFeatureTablesFromGeopackage()
         {
             InitializeComponent();
             _ = Initialize();
@@ -33,21 +35,18 @@ namespace ArcGIS.Samples.ReadGeoPackage
         {
             // Create a new map centered on Aurora Colorado.
             MyMapView.Map = new Map(BasemapStyle.ArcGISStreets);
-            MyMapView.Map.InitialViewpoint = new Viewpoint(39.7294, -104.70, 175000);
+            MyMapView.Map.InitialViewpoint = new Viewpoint(39.7294, -104.73, 175000);
 
             // Get the full path to the GeoPackage on the device.
-            string myGeoPackagePath = GetGeoPackagePath();
+            string myGeoPackagePath = DataManager.GetDataFolder("68ec42517cdd439e81b036210483e8e7", "AuroraCO.gpkg");
 
             try
             {
                 // Open the GeoPackage.
                 GeoPackage myGeoPackage = await GeoPackage.OpenAsync(myGeoPackagePath);
 
-                // Get the read only list of GeoPackageRasters from the GeoPackage.
-                IReadOnlyList<GeoPackageRaster> myReadOnlyListOfGeoPackageRasters = myGeoPackage.GeoPackageRasters;
-
                 // Loop through each GeoPackageRaster.
-                foreach (GeoPackageRaster oneGeoPackageRaster in myReadOnlyListOfGeoPackageRasters)
+                foreach (GeoPackageRaster oneGeoPackageRaster in myGeoPackage.GeoPackageRasters)
                 {
                     // Create a RasterLayer from the GeoPackageRaster.
                     RasterLayer myRasterLayer = new RasterLayer(oneGeoPackageRaster)
@@ -56,15 +55,12 @@ namespace ArcGIS.Samples.ReadGeoPackage
                         Opacity = 0.55
                     };
 
-                    // Add the layer to the map.
+                    // Add the raster layer to the map.
                     MyMapView.Map.OperationalLayers.Add(myRasterLayer);
                 }
 
-                // Get the read only list of GeoPackageFeatureTables from the GeoPackage.
-                IReadOnlyList<GeoPackageFeatureTable> myReadOnlyListOfGeoPackageFeatureTables = myGeoPackage.GeoPackageFeatureTables;
-
                 // Loop through each GeoPackageFeatureTable.
-                foreach (GeoPackageFeatureTable oneGeoPackageFeatureTable in myReadOnlyListOfGeoPackageFeatureTables)
+                foreach (GeoPackageFeatureTable oneGeoPackageFeatureTable in myGeoPackage.GeoPackageFeatureTables)
                 {
                     // Create a FeatureLayer from the GeoPackageFeatureLayer.
                     FeatureLayer myFeatureLayer = new FeatureLayer(oneGeoPackageFeatureTable);
@@ -75,10 +71,8 @@ namespace ArcGIS.Samples.ReadGeoPackage
             }
             catch (Exception e)
             {
-                await Application.Current.Windows[0].Page.DisplayAlert("Error", e.ToString(), "OK");
+                await new MessageDialog2(e.ToString(), "Error").ShowAsync();
             }
         }
-
-        private static string GetGeoPackagePath() => DataManager.GetDataFolder("68ec42517cdd439e81b036210483e8e7", "AuroraCO.gpkg");
     }
 }
