--- src/WinUI/ArcGIS.WinUI.Viewer/Samples/Data/RasterLayerGeoPackage/RasterLayerGeoPackage.xaml
@@ -1,6 +1,0 @@
-﻿<UserControl x:Class="ArcGIS.WinUI.Samples.RasterLayerGeoPackage.RasterLayerGeoPackage"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:esriUI="using:Esri.ArcGISRuntime.UI.Controls">
-    <esriUI:MapView x:Name="MyMapView" />
-</UserControl>
